+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Releases Changelog
 
+## 1.6.0
+- Added support for Kerberos based authentication DBs. In order to enable it on a exising or new dataset, a secret must be created beforehand in the same namespace where the federated node is deployed.
+
+    This secret, should have two keys: `krb5.conf` with the file contents of the omonimous file, and `principal.keytab` with the file contents of the principal keytab.
+
+    And example from CLI would be:
+    ```sh
+    kubectl create secret -n federatednode generic kerb-db-auth --from-file=krb5.conf=/etc/krb5.conf --from-file=principal.keytab=/data/mssql.keytab
+    ```
+
+    Once this is created, the POST/PATCH endpoints can be used to map the dataset with this type of auth, using the following as integration to the request body:
+    ```json
+    {
+        "auth_type": "kerberos",
+        "kerberos": {
+            "secret_name": "kerb-db-auth"
+        }
+    }
+    ```
+
 ## 1.5.0
 - Prefixed cluster-wide resources with the release name (unique by helm standards). Moved unnecessarily cluster-wide resources to namespaced ones
 - Added the option to setup an initial user to avoid using the backend credentials. To set it up, the following section in the values file has been added: